TitanBot/requirements.txt
# CORE INFRASTRUCTURE
aiohttp==3.9.3
aiosqlite==0.20.0
uvloop==0.19.0; sys_platform != 'win32'
python-dotenv==1.0.1
colorlog==6.8.2

# DATA SCIENCE & MATH
numpy==1.26.4
pandas==2.2.1
scipy==1.12.0
ta-lib==0.4.28

# CRYPTO & API
ccxt==4.2.19
python-telegram-bot==20.8

# AI/ML (OPTIONAL - for sentiment analysis)
google-genai>=0.3.0  # NEW Gemini library
# google-generativeai>=0.3.0  # OLD (deprecated) - uncomment if needed

# UTILITIES
pyyaml==6.0.1